refactor: moved shared folders to specific tab component
