+++ src/app/divider/shared/card-type.model.ts
