+++ src/app/gearbox/shared/ship.model.ts
